add allocator e2e test
