+++ integration-tests/static-buffer/lib.rs
@@ -2,6 +2,8 @@
 
 #[ink::contract]
 pub mod static_buffer {
+    use ink::prelude::vec::Vec;
+
     #[allow(unused_imports)]
     use ink::env::BUFFER_SIZE;
     #[ink(storage)]
@@ -27,6 +29,15 @@ pub mod static_buffer {
         #[ink(message)]
         pub fn get_caller(&self) -> AccountId {
             self.env().caller()
+        }
+
+        #[ink(message)]
+        pub fn buffer(&self) {
+            let buf1 = Vec::<u8>::with_capacity(3);
+            let buf2 = Vec::<u64>::with_capacity(1);
+            ink::env::debug_println!("{:?}", buf1.as_ptr());
+            ink::env::debug_println!("{:?}", buf2.as_ptr());
+            ink::env::debug_println!("{}", core::mem::align_of::<Vec<bool>>());
         }
     }
 
@@ -72,6 +83,35 @@ pub mod static_buffer {
                 "Buffer size was larger than expected: {}",
                 super::BUFFER_SIZE.to_string()
             );
+
+            Ok(())
+        }
+
+        #[ink_e2e::test]
+        async fn buffer<Client: E2EBackend>(mut client: Client) -> E2EResult<()> {
+            // given
+            let mut constructor = StaticBufferRef::new_default();
+
+            // when
+            let contract = client
+                .instantiate("static_buffer", &ink_e2e::bob(), &mut constructor)
+                .submit()
+                .await
+                .expect("instantiate failed");
+            let call_builder = contract.call_builder::<StaticBuffer>();
+
+            // then
+            let get = call_builder.buffer();
+            let get_res = client.call(&ink_e2e::bob(), &get).submit().await?;
+
+            let debug_msg = get_res.debug_message();
+            let msgs: Vec<&str> = debug_msg.split('\n').collect();
+            let ptr1 = u64::from_str_radix(msgs[0].trim_start_matches("0x"), 16).unwrap();
+            let ptr2 = u64::from_str_radix(msgs[1].trim_start_matches("0x"), 16).unwrap();
+            let align = u64::from_str_radix(msgs[2], 10).unwrap();
+
+            assert_eq!(align, 4);
+            assert_eq!((ptr2 - ptr1), 8);
 
             Ok(())
         }
